+++ Source/BlueprintAssist/Private/BlueprintAssistCache.cpp
@@ -263,6 +263,18 @@ void FBACache::SaveGraphDataToPackageMetaData(UEdGraph* Graph)
 
 	if (UPackage* AssetPackage = Graph->GetPackage())
 	{
+#if ENGINE_MAJOR_VERSION >= 5 && ENGINE_MINOR_VERSION >= 6
+		FMetaData& MetaData = AssetPackage->GetMetaData();
+		FBAGraphData& GraphData = GetGraphData(Graph);
+
+		GraphData.CleanupGraph(Graph);
+		
+		FString GraphDataAsString;
+		if (FJsonObjectConverter::UStructToJsonObjectString(GraphData, GraphDataAsString))
+		{
+			MetaData.SetValue(Graph, NAME_BA_GRAPH_DATA, *GraphDataAsString);
+		}
+#else
 		if (UMetaData* MetaData = AssetPackage->GetMetaData())
 		{
 			FBAGraphData& GraphData = GetGraphData(Graph);
@@ -275,6 +287,7 @@ void FBACache::SaveGraphDataToPackageMetaData(UEdGraph* Graph)
 				MetaData->SetValue(Graph, NAME_BA_GRAPH_DATA, *GraphDataAsString);
 			}
 		}
+#endif
 	}
 }
 
@@ -292,6 +305,17 @@ bool FBACache::LoadGraphDataFromPackageMetaData(UEdGraph* Graph, FBAGraphData& G
 
 	if (UPackage* AssetPackage = Graph->GetPackage())
 	{
+#if ENGINE_MAJOR_VERSION >= 5 && ENGINE_MINOR_VERSION >= 6
+		FMetaData& MetaData = AssetPackage->GetMetaData();
+		if (const FString* GraphDataAsString = MetaData.FindValue(Graph, NAME_BA_GRAPH_DATA))
+		{
+			if (FJsonObjectConverter::JsonObjectStringToUStruct(*GraphDataAsString, &GraphData, 0, 0))
+			{
+				GraphData.bTriedLoadingMetaData = true;
+				return true;
+			}
+		}
+#else
 		if (UMetaData* MetaData = AssetPackage->GetMetaData())
 		{
 			if (const FString* GraphDataAsString = MetaData->FindValue(Graph, NAME_BA_GRAPH_DATA))
@@ -303,6 +327,7 @@ bool FBACache::LoadGraphDataFromPackageMetaData(UEdGraph* Graph, FBAGraphData& G
 				}
 			}
 		}
+#endif
 	}
 
 	return false;
@@ -312,10 +337,15 @@ void FBACache::ClearPackageMetaData(UEdGraph* Graph)
 {
 	if (UPackage* AssetPackage = Graph->GetPackage())
 	{
+#if ENGINE_MAJOR_VERSION >= 5 && ENGINE_MINOR_VERSION >= 6
+		FMetaData& MetaData = AssetPackage->GetMetaData();
+		MetaData.RemoveValue(Graph, NAME_BA_GRAPH_DATA);
+#else
 		if (UMetaData* MetaData = AssetPackage->GetMetaData())
 		{
 			MetaData->RemoveValue(Graph, NAME_BA_GRAPH_DATA);
 		}
+#endif
 	}
 }